requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
numpy>=1.26.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
